Don't send duplicate Date: header
Asterisk writes the "Date:" header into email bdy, but s-nail writes it again. It isn't possible to disable it from configuration in voicemail.conf or s-nail.rc, therefore here there is a simple bash script that removes the Date: header then pipes the mail to s-nail. It is possible to improve the script further in future, for instance by sending the voicemail to other channels or do STT on the attachment.