clone-module: preserve new passwords file
The cloning procedure rewrites the newly generated password file with
the old one from the origin module, causing the new passwords to be lost
and the old ones to be used.

To prevent this issue, the new password file is preserved by renaming it
to `passwords.new` and adding an rsync filter rule in
`etc/state-filter-rules.rsync` to exclude it from deletion. This file is
then used to reset the old passwords.