feat(RPS): update RPS url on migration or when hostname changes
On freepbx container start, NETHVOICE_HOST saved into db is checked. If it exists and is different from environment NETHVOICE_HOST, a new tok1 is generated for the phones and new configuration url is putted in RPS
If the phone is resetted, it will use new url and start again a new provisioning.

- On migration, the NETHVOICE_HOST variable is created, so it differs from environment and RPS is updated
- On normal update, NETHVOICE_HOST variable doesn't exist and RPS is not updated
- On NETHVOICE_HOST changed, RPS is updated.

This requires Tancredi API for updating tok1