Documents extension ring script
Adds documentation for the `NETHCTI_CDR_SCRIPT_EXTENSION_RING`
option, which allows a script to be executed when an extension rings.

This enhancement enables greater flexibility in call management and
integration with external systems.

Adds called extension parameter

Exposes the called extension as a parameter in the extension rings script.

This allows scripts to make decisions based on the specific extension that triggered the ring event,
enabling more flexible and targeted call handling.